+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/UnitTests.hs
@@ -73,17 +73,17 @@ import Cardano.Ledger.Shelley.TxBody
   ( PoolMetadata (..),
     PoolParams (..),
     Wdrl (..),
-    _poolCost,
-    _poolId,
-    _poolMD,
-    _poolMDHash,
-    _poolMDUrl,
-    _poolMargin,
-    _poolOwners,
-    _poolPledge,
-    _poolRAcnt,
-    _poolRelays,
-    _poolVrf,
+    pmHash,
+    pmUrl,
+    ppCost,
+    ppId,
+    ppMargin,
+    ppMetadata,
+    ppOwners,
+    ppPledge,
+    ppRelays,
+    ppRewardAcnt,
+    ppVrf,
     pattern RewardAcnt,
   )
 import Cardano.Ledger.Shelley.TxWits
@@ -666,19 +666,19 @@ alicePoolColdKeys = KeyPair vk sk
 alicePoolParamsSmallCost :: PoolParams C_Crypto
 alicePoolParamsSmallCost =
   PoolParams
-    { _poolId = hashKey . vKey $ alicePoolColdKeys,
-      _poolVrf = hashVerKeyVRF vkVrf,
-      _poolPledge = Coin 1,
-      _poolCost = Coin 5, -- Too Small!
-      _poolMargin = unsafeBoundRational 0.1,
-      _poolRAcnt = RewardAcnt Testnet (KeyHashObj . hashKey . vKey $ aliceStake),
-      _poolOwners = Set.singleton $ (hashKey . vKey) aliceStake,
-      _poolRelays = StrictSeq.empty,
-      _poolMD =
+    { ppId = hashKey . vKey $ alicePoolColdKeys,
+      ppVrf = hashVerKeyVRF vkVrf,
+      ppPledge = Coin 1,
+      ppCost = Coin 5, -- Too Small!
+      ppMargin = unsafeBoundRational 0.1,
+      ppRewardAcnt = RewardAcnt Testnet (KeyHashObj . hashKey . vKey $ aliceStake),
+      ppOwners = Set.singleton $ (hashKey . vKey) aliceStake,
+      ppRelays = StrictSeq.empty,
+      ppMetadata =
         SJust $
           PoolMetadata
-            { _poolMDUrl = fromJust $ textToUrl "alice.pool",
-              _poolMDHash = BS.pack "{}"
+            { pmUrl = fromJust $ textToUrl "alice.pool",
+              pmHash = BS.pack "{}"
             }
     }
   where
@@ -690,7 +690,7 @@ testPoolCostTooSmall =
     [ DelegsFailure
         ( DelplFailure
             ( PoolFailure
-                ( StakePoolCostTooLowPOOL (_poolCost alicePoolParamsSmallCost) (_minPoolCost (pp @C))
+                ( StakePoolCostTooLowPOOL (ppCost alicePoolParamsSmallCost) (_minPoolCost (pp @C))
                 )
             )
         )